+++ src/database/queries/proposals/insertOne.sql
@@ -1,8 +1,10 @@
 INSERT INTO proposals (
   external_id,
-  opportunity_id
+  opportunity_id,
+  created_by
 ) VALUES (
   :externalId,
-  :opportunityId
+  :opportunityId,
+  :createdBy
 )
 RETURNING proposal_to_json(proposals) AS "object";