+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The `BulkUpload` entity now has a `createdBy` attribute.
+- The `Proposal` entity now has a `createdBy` attribute.
 
 ## 0.7.0 2024-04-04